refactored customer and filter service
