+++ frontend/frontendAngular/src/app/services/filter.service.ts
@@ -1,20 +1,32 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, of} from 'rxjs';
 import { CompanyModel } from '../api/company';
 import { LocalStorageLoginService } from './local-storage-login.service';
 
-const apiUrl = 'http://127.0.0.1:8000/api/auth/companies/filter';
+const apiUrl = 'http://127.0.0.1:8000/api/auth/companies';
 
 @Injectable({
   providedIn: 'root'
 })
-export class CategoryFilterService {
+export class FilterService {
 
   constructor(
     private http: HttpClient,
     public localStorage: LocalStorageLoginService,
   ) { }
+
+  search(term: string): Observable<CompanyModel[]> {
+    const token = this.localStorage.get("token");
+    const header = new HttpHeaders({ 'Authorization': `Bearer ${token}` });
+    const options = {
+      headers: header,
+    };
+    if (!term.trim()) {
+      return of([]);
+    }
+    return this.http.get<CompanyModel[]>(`${apiUrl}/search?search=${term}`, options)
+  }
 
   filter(term: string): Observable<CompanyModel[]> {
     const token = this.localStorage.get("token");
@@ -22,6 +34,6 @@ export class CategoryFilterService {
     const options = {
       headers: header,
     };
-    return this.http.get<CompanyModel[]>(`${apiUrl}?category=${term}`, options)
+    return this.http.get<CompanyModel[]>(`${apiUrl}/filter?category=${term}`, options)
   }
 }
